removed log4cpp, fixed a bug in librexgen_lua.cpp, introduced install target
git-svn-id: http://rexgen.googlecode.com/svn/trunk@12 c03ab62d-064e-604f-90e8-bc6f304d3419

+++ src/librexgen/iterator/iteratorpermuter.h
@@ -153,7 +153,7 @@ void IteratorPermuter<REGEX>::next()
   
   assert(canUseValue());
   
-  TRACE_INT("next2: current length: %d", iterators.size());
+  TRACE_INT("next2: current length: %d", (int)iterators.size());
   LEAVE_METHOD;
 }
